Reestrucura Makefile y implementamos control_de_parametros
